Security hardening, test infrastructure, and cleanup
- Add CSRF middleware with multipart exemptions for file upload routes
- Add account lockout (5 attempts, 15 min) and TOCTOU fix in auth controller
- Add audit log sensitive field redaction
- Add template injection prevention and cross-tenant workflow scoping
- Add transaction wrapping for campaign analytics tracking
- Add env var validation, per-route body size limits, graceful cron shutdown
- Strengthen password validation requirements
- Add performance indexes and user address/lockout fields to schema
- Fix GlobalSearchModal URL routing (/automation -> /workflows)
- Add duplicate field key detection in CustomFieldsManager
- Add accessibility improvements to NotificationBell
- Fix ToastContainer z-index for modal overlay
- Improve NodeConfigPanel template variable display
- Add analytics filters (source, status, priority), lead trends, time-to-convert
- Add 441 backend tests (middleware, services, utils) with factory helpers
- Add 260 frontend tests (components, hooks, stores, utils)
- Add e2e test scaffolding (auth, navigation specs)
- Add API reference, architecture, and setup guide docs
- Clean up 77 outdated audit/test/backup files
- Harden various page components (pagination guards, error handling, XSS prevention)

+++ src/pages/admin/DatabaseMaintenance.tsx
@@ -255,7 +255,7 @@ const DatabaseMaintenance = () => {
                       <div><span className="text-muted-foreground">Engine:</span> <strong>PostgreSQL</strong></div>
                     </div>
                     <div className="grid grid-cols-2 md:grid-cols-4 gap-3">
-                      <div><span className="text-muted-foreground">Index Size:</span> <strong>~{(parseInt(table.size) * 0.15).toFixed(0) || '18'} MB</strong></div>
+                      <div><span className="text-muted-foreground">Index Size:</span> <strong>~{(parseFloat(table.size) * 0.15).toFixed(0) || '18'} MB</strong></div>
                       <div><span className="text-muted-foreground">Avg Row Size:</span> <strong>{table.records > 0 ? Math.round((parseFloat(table.size) * 1024 * 1024) / table.records) : 0} bytes</strong></div>
                       <div><span className="text-muted-foreground">Dead Tuples:</span> <strong>~{Math.round(table.records * 0.012).toLocaleString()}</strong></div>
                       <div><span className="text-muted-foreground">Status:</span> <Badge variant="default">Healthy</Badge></div>
